checkpoint with student data insert and delete
complete working of insert and delete of students data

+++ src/main/java/com/example/StudentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "student", path = "student")
-public interface StudentRepository extends PagingAndSortingRepository<Student, Long> {
+public interface StudentRepository extends PagingAndSortingRepository<Student, Integer> {
 
 	List<Student> findByLastName(@Param("name") String name);
